refactor(facturacion): reorganizar proceso de facturación y actualizar dependencias
* feat: crear FacturacionController y DatoUnaFacturaDto
* feat: implementar FacturacionService y ContabilidadService
* refactor: simplificar MakeFacturaProgramaDiaService
* chore: actualizar dependencias del proyecto
  - spring-boot 3.4.4
  - kotlin 2.1.20
  - spring-cloud 2024.0.1
  - springdoc-openapi 2.8.6
* chore: eliminar VentaDto obsoleto
* chore: actualizar configuración CORS

La refactorización mejora la organización del código separando las responsabilidades
en servicios específicos para facturación y contabilidad. Se elimina código duplicado
y se mejora el manejo de nulos y logging.

Closes #83
