+++ src/main/java/eterea/core/service/model/dto/DatoUnaFacturaDto.java
@@ -1,0 +1,13 @@
+package eterea.core.service.model.dto;
+
+import eterea.core.service.kotlin.model.ArticuloMovimiento;
+import eterea.core.service.kotlin.model.ClienteMovimiento;
+import lombok.Data;
+
+@Data
+public class DatoUnaFacturaDto {
+
+        private ClienteMovimiento clienteMovimiento;
+        private ArticuloMovimiento articuloMovimiento;
+
+}
